+++ public/index.html
@@ -14,61 +14,34 @@
 <header class="header">
   <h1 class="header__logo logo">Cinemaddict</h1>
 
-  <!-- Звание пользователя -->
+
 </header>
 
 <main class="main">
-  <!-- Меню -->
-
-  <!-- Сортировка -->
-
-  <!-- Контент -->
   <section class="films">
     <section class="films-list">
       <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-
       <div class="films-list__container">
-
       </div>
-
     </section>
-
-
-
     <section class="films-list--extra">
       <h2 class="films-list__title">Top rated</h2>
-
       <div class="films-list__container">
-
           </div>
     </section>
-
     <section class="films-list--extra">
       <h2 class="films-list__title">Most commented</h2>
-
       <div class="films-list__container films-list__container--second">
-
-
-
       </div>
     </section>
   </section>
 </main>
-
-
-
-
-
-
 </main>
-
-
 <footer class="footer">
   <section class="footer__logo logo logo--smaller">Cinemaddict</section>
   <section class="footer__statistics">
-    <!-- Количество фильмов -->
   </section>
 </footer>
-<script src="../src/main.js"></script>
+<script src="./bundle.js"></script>
 </body>
 </html>
