+++ src/view/buttonShowMore.js
@@ -1,0 +1,7 @@
+const createButtonShowMore = () => {
+    return (
+      `<button class="films-list__show-more">Show more</button>`
+    );
+  };
+
+  export { createButtonShowMore };
